Backend/requirements.txt
flask
gunicorn
flask-cors
pillow
numpy
requests
flask-restful
python-dotenv
opencv-python

